+++ src/components/manage/Form/Form.jsx
@@ -53,6 +53,18 @@ const messages = defineMessages({
     id: 'Minimum length is {len}.',
     defaultMessage: 'Minimum length is {len}.',
   },
+  maxLength: {
+    id: 'Maximum length is {len}.',
+    defaultMessage: 'Maximum length is {len}.',
+  },
+  minimum: {
+    id: 'Minimum value is {len}.',
+    defaultMessage: 'Minimum value is {len}.',
+  },
+  maximum: {
+    id: 'Maximum value is {len}.',
+    defaultMessage: 'Maximum value is {len}.',
+  },
   uniqueItems: {
     id: 'Items must be unique.',
     defaultMessage: 'Items must be unique.',
@@ -60,6 +72,23 @@ const messages = defineMessages({
   save: {
     id: 'Save',
     defaultMessage: 'Save',
+  },
+  isNumber: {
+    id: 'number',
+    defaultMessage: 'Input must be number',
+  },
+  isInteger: {
+    id: 'integer',
+    defaultMessage: 'Input must be integer',
+  },
+  isValidEmail: {
+    id: 'email',
+    defaultMessage: 'Input must be valid email (something@domain.com)',
+  },
+  isValidURL: {
+    id: 'url',
+    defaultMessage:
+      'Input must be valid url (www.something.com or http(s)://www.something.com)',
   },
   cancel: {
     id: 'Cancel',
@@ -74,6 +103,159 @@ const messages = defineMessages({
     defaultMessage: 'There were some errors.',
   },
 });
+
+/**
+ * Will return the intl message if invalid
+ * @param {boolean} isValid
+ * @param {string} maxCriterion
+ * @param {string | number} valueToCompare can compare '47' < 50
+ * @param {Function} intlFunc
+ */
+const validationMessage = (isValid, maxCriterion, valueToCompare, intlFunc) =>
+  !isValid
+    ? intlFunc(messages[maxCriterion], {
+        len: valueToCompare,
+      })
+    : null;
+/**
+ * Returns if based on the criterion the value is lower or equal
+ * @param {string | number} value can compare '47' < 50
+ * @param {string | number} valueToCompare can compare '47' < 50
+ * @param {string} minCriterion
+ * @param {Function} intlFunc
+ */
+const isMaxPropertyValid = (value, valueToCompare, minCriterion, intlFunc) => {
+  const isValid = valueToCompare !== undefined ? value <= valueToCompare : true;
+  return validationMessage(isValid, minCriterion, valueToCompare, intlFunc);
+};
+/**
+ * Returns if based on the criterion the value is higher or equal
+ * @param {string | number} value can compare '47' < 50
+ * @param {string | number} valueToCompare can compare '47' < 50
+ * @param {string} minCriterion
+ * @param {Function} intlFunc
+ */
+const isMinPropertyValid = (value, valueToCompare, maxCriterion, intlFunc) => {
+  const isValid = valueToCompare !== undefined ? value >= valueToCompare : true;
+  return validationMessage(isValid, maxCriterion, valueToCompare, intlFunc);
+};
+
+const widgetValidation = {
+  email: {
+    isValidEmail: (emailValue, emailObj, intlFunc) => {
+      const emailRegex = /^\w+([.-]?\w+)*@\w+([.-]?\w+)*(\.\w{2,3})+$/;
+      const isValid = emailRegex.test(emailValue);
+      return !isValid ? intlFunc(messages.isValidEmail) : null;
+    },
+    minLength: (emailValue, emailObj, intlFunc) =>
+      isMinPropertyValid(
+        emailValue.length,
+        emailObj.minLength,
+        'minLength',
+        intlFunc,
+      ),
+    maxLength: (emailValue, emailObj, intlFunc) =>
+      isMaxPropertyValid(
+        emailValue.length,
+        emailObj.maxLength,
+        'maxLength',
+        intlFunc,
+      ),
+  },
+  url: {
+    isValidURL: (urlValue, urlObj, intlFunc) => {
+      const urlRegex = /^(http:\/\/www\.|https:\/\/www\.|http:\/\/|https:\/\/)?[a-z0-9]+([-.]{1}[a-z0-9]+)*\.[a-z]{2,5}(:[0-9]{1,5})?(\/.*)?$/gm;
+      const isValid = urlRegex.test(urlValue);
+      return !isValid ? intlFunc(messages.isValidURL) : null;
+    },
+    minLength: (urlValue, urlObj, intlFunc) =>
+      isMinPropertyValid(
+        urlValue.length,
+        urlObj.minLength,
+        'minLength',
+        intlFunc,
+      ),
+    maxLength: (urlValue, urlObj, intlFunc) =>
+      isMaxPropertyValid(
+        urlValue.length,
+        urlObj.maxLength,
+        'maxLength',
+        intlFunc,
+      ),
+  },
+  password: {
+    minLength: (passwordValue, passwordObj, intlFunc) =>
+      isMinPropertyValid(
+        passwordValue.length,
+        passwordObj.minLength,
+        'minLength',
+        intlFunc,
+      ),
+    maxLength: (passwordValue, passwordObj, intlFunc) =>
+      isMaxPropertyValid(
+        passwordValue.length,
+        passwordObj.maxLength,
+        'maxLength',
+        intlFunc,
+      ),
+  },
+  string: {
+    minLength: (value, itemObj, intlFunc) =>
+      isMinPropertyValid(
+        value.length,
+        itemObj.minLength,
+        'minLength',
+        intlFunc,
+      ),
+    maxLength: (value, itemObj, intlFunc) =>
+      isMaxPropertyValid(
+        value.length,
+        itemObj.maxLengthj,
+        'maxLength',
+        intlFunc,
+      ),
+  },
+  number: {
+    isNumber: (value, itemObj, intlFunc) => {
+      const floatRegex = /^[+-]?\d+(\.\d+)?$/;
+      const isValid = !isNaN(value) && floatRegex.test(value);
+      return !isValid ? intlFunc(messages.isNumber) : null;
+    },
+    minimum: (value, itemObj, intlFunc) =>
+      isMinPropertyValid(value, itemObj.minimum, 'minimum', intlFunc),
+    maximum: (value, itemObj, intlFunc) =>
+      isMaxPropertyValid(value, itemObj.maximum, 'maximum', intlFunc),
+  },
+  integer: {
+    isInteger: (value, itemObj, intlFunc) => {
+      const intRegex = /^-?[0-9]+$/;
+      const isValid = !isNaN(value) && intRegex.test(value);
+      return !isValid ? intlFunc(messages.isInteger) : null;
+    },
+    minimum: (value, itemObj, intlFunc) =>
+      isMinPropertyValid(value, itemObj.minimum, 'minimum', intlFunc),
+    maximum: (value, itemObj, intlFunc) =>
+      isMaxPropertyValid(value, itemObj.maximum, 'maximum', intlFunc),
+  },
+};
+
+/**
+ * The string that comes my not be a valid JSON
+ * @param {string} requestItem
+ */
+const tryParseJSON = requestItem => {
+  let resultObj = null;
+  try {
+    resultObj = JSON.parse(requestItem);
+  } catch (e) {
+    try {
+      resultObj = JSON.parse(requestItem.replace(/'/g, '"'));
+    } catch (e) {
+      resultObj = null;
+    }
+  }
+  return resultObj;
+};
 
 /**
  * Form container class.
@@ -115,6 +297,7 @@ class Form extends Component {
     description: PropTypes.string,
     visual: PropTypes.bool,
     blocks: PropTypes.arrayOf(PropTypes.object),
+    requestError: PropTypes.string,
   };
 
   /**
@@ -138,6 +321,7 @@ class Form extends Component {
     blocks: [],
     pathname: '',
     schema: {},
+    requestError: null,
   };
 
   /**
@@ -187,13 +371,17 @@ class Form extends Component {
     this.state = {
       formData,
       initialFormData: { ...formData },
+      isFormPrestine: true,
       errors: {},
+      activeIndex: 0,
       selected:
         formData[blocksLayoutFieldname].items.length > 0
           ? formData[blocksLayoutFieldname].items[0]
           : null,
     };
     this.onChangeField = this.onChangeField.bind(this);
+    this.onBlurField = this.onBlurField.bind(this);
+    this.onClickInput = this.onClickInput.bind(this);
     this.onChangeBlock = this.onChangeBlock.bind(this);
     this.onMutateBlock = this.onMutateBlock.bind(this);
     this.onSelectBlock = this.onSelectBlock.bind(this);
@@ -204,22 +392,126 @@ class Form extends Component {
     this.onFocusPreviousBlock = this.onFocusPreviousBlock.bind(this);
     this.onFocusNextBlock = this.onFocusNextBlock.bind(this);
     this.handleKeyDown = this.handleKeyDown.bind(this);
+    this.onTabChange = this.onTabChange.bind(this);
+  }
+
+  /**
+   * The first Fieldset (Tab) that has any errors
+   * will be selected
+   * @param {Object[]} errors
+   * @param {string} errors[].field
+   * @param {string[]} errors[].message
+   * @returns {number} activeIndex
+   */
+  showFirstTabWithErrors(errors) {
+    let activeIndex = 0;
+
+    this.props.schema.fieldsets.some((fieldSet, index) => {
+      let foundfield = fieldSet.fields.some(fieldId => errors[fieldId]);
+
+      activeIndex = foundfield ? index : activeIndex;
+      return foundfield;
+    });
+
+    return activeIndex;
+  }
+
+  /**
+   * Create the errors object from backend the same way it is done on Frontend validation
+   * @param {string} requestError form the server
+   * @returns {Object}
+   */
+  giveServerErrorsToCorrespondingFields(requestError) {
+    let errorsList = tryParseJSON(requestError);
+    const errors = {};
+
+    if (Array.isArray(errorsList) && errorsList.length > 0) {
+      errorsList.forEach(errorItem => {
+        errors[errorItem.field] = errors[errorItem.field]
+          ? errors[errorItem.field].push(errorItem.message)
+          : [errorItem.message];
+      });
+    }
+    return errors;
+  }
+
+  /**
+   * On updates caused by props change
+   * if errors from Backend come, these will be shown to their corresponding Fields
+   * also the first Tab to have any errors will be selected
+   * @param {Object} prevProps
+   */
+  componentDidUpdate(prevProps) {
+    let { requestError } = this.props;
+    let errors = {};
+    let activeIndex = 0;
+
+    if (requestError && prevProps.requestError !== requestError) {
+      errors = this.giveServerErrorsToCorrespondingFields(requestError);
+      activeIndex = this.showFirstTabWithErrors(errors);
+
+      this.setState({
+        errors,
+        activeIndex,
+      });
+    }
+  }
+
+  /**
+   * Tab selection is done only by setting activeIndex in state
+   */
+  onTabChange(e, { activeIndex }) {
+    this.setState({ activeIndex });
+  }
+
+  /**
+   * If user clicks on input, the form will be not considered pristine
+   * this will avoid onBlur effects without interraction with the form
+   * @param {Object} e event
+   */
+  onClickInput(e) {
+    this.setState({ isFormPrestine: false });
+  }
+
+  /**
+   * Validate fields on blur
+   * @method onBlurField
+   * @param {string} id Id of the field
+   * @param {*} value Value of the field
+   * @returns {undefined}
+   */
+  onBlurField(id, value) {
+    if (!this.state.isFormPrestine) {
+      const errors = this.validateFieldsPerFieldset();
+      if (keys(errors).length > 0) {
+        this.setState({
+          errors,
+        });
+      }
+    }
   }
 
   /**
    * Change field handler
+   * Remove errors for changed field
    * @method onChangeField
    * @param {string} id Id of the field
    * @param {*} value Value of the field
    * @returns {undefined}
    */
   onChangeField(id, value) {
-    this.setState({
-      formData: {
-        ...this.state.formData,
-        // We need to catch also when the value equals false this fixes #888
-        [id]: value || (value !== undefined && isBoolean(value)) ? value : null,
-      },
+    this.setState(prevState => {
+      const { errors, formData } = prevState;
+      delete errors[id];
+      return {
+        errors,
+        formData: {
+          ...formData,
+          // We need to catch also when the value equals false this fixes #888
+          [id]:
+            value || (value !== undefined && isBoolean(value)) ? value : null,
+        },
+      };
     });
   }
 
@@ -365,7 +657,90 @@ class Form extends Component {
   }
 
   /**
-   * Submit handler
+   * Returns errors if obj has unique Items
+   * @param {Object} field
+   * @param {*} fieldData
+   * @returns {Object[string]} - list of errors
+   */
+  hasUniqueItems(field, fieldData) {
+    const errors = [];
+    if (
+      field.uniqueItems &&
+      fieldData &&
+      uniq(fieldData).length !== fieldData.length
+    ) {
+      errors.push(this.props.intl.formatMessage(messages.uniqueItems));
+    }
+    return errors;
+  }
+
+  /**
+   * If required fields are undefined, return list of errors
+   * @returns {Object[string]} - list of errors
+   */
+  validateRequiredFields() {
+    const errors = {};
+
+    map(this.props.schema.required, requiredField => {
+      if (!this.state.formData[requiredField]) {
+        errors[requiredField] = [];
+        errors[requiredField].push(
+          this.props.intl.formatMessage(messages.required),
+        );
+      }
+    });
+
+    return errors;
+  }
+
+  /**
+   * Return list of errors if field constraints are not respected
+   * (ex min, max, maxLength, email format, url format etc)
+   * each potential criterion has a validation process in widgetValidation
+   * !!ONLY fields with data will be tested (those undefined are ignored here)
+   * @returns {Object[string]} - list of errors
+   */
+  validateFieldsPerFieldset() {
+    const errors = this.validateRequiredFields();
+
+    map(this.props.schema.properties, (field, fieldId) => {
+      const fieldWidgetType = field.widget || field.type;
+      const widgetValidationCriteria = widgetValidation[fieldWidgetType]
+        ? Object.keys(widgetValidation[fieldWidgetType])
+        : [];
+      let fieldData = this.state.formData[fieldId];
+      // test each criterion ex maximum, isEmail, isUrl, maxLength etc
+      const fieldErrors = widgetValidationCriteria
+        .map(widgetCriterion => {
+          const errorMessage =
+            fieldData === undefined || fieldData === null
+              ? null
+              : widgetValidation[fieldWidgetType][widgetCriterion](
+                  fieldData,
+                  field,
+                  this.props.intl.formatMessage,
+                );
+          return errorMessage;
+        })
+        .filter(item => !!item);
+
+      const uniqueErrors = this.hasUniqueItems(field, fieldData);
+      const mergedErrors = [...fieldErrors, ...uniqueErrors];
+
+      if (mergedErrors.length > 0) {
+        errors[fieldId] = [
+          ...(errors[fieldId] || []),
+          ...fieldErrors,
+          ...uniqueErrors,
+        ];
+      }
+    });
+
+    return errors;
+  }
+
+  /**
+   * Submit handler also validate form and collect errors
    * @method onSubmit
    * @param {Object} event Event object.
    * @returns {undefined}
@@ -374,38 +749,9 @@ class Form extends Component {
     if (event) {
       event.preventDefault();
     }
-    const errors = {};
-    map(this.props.schema.fieldsets, fieldset =>
-      map(fieldset.fields, fieldId => {
-        const field = this.props.schema.properties[fieldId];
-        var data = this.state.formData[fieldId];
-        if (typeof data === 'string' || data instanceof String) {
-          data = data.trim();
-        }
-        if (this.props.schema.required.indexOf(fieldId) !== -1) {
-          if (field.type !== 'boolean' && !data) {
-            errors[fieldId] = errors[field] || [];
-            errors[fieldId].push(
-              this.props.intl.formatMessage(messages.required),
-            );
-          }
-          if (field.minLength && data.length < field.minLength) {
-            errors[fieldId] = errors[field] || [];
-            errors[fieldId].push(
-              this.props.intl.formatMessage(messages.minLength, {
-                len: field.minLength,
-              }),
-            );
-          }
-        }
-        if (field.uniqueItems && data && uniq(data).length !== data.length) {
-          errors[fieldId] = errors[field] || [];
-          errors[fieldId].push(
-            this.props.intl.formatMessage(messages.uniqueItems),
-          );
-        }
-      }),
-    );
+
+    const errors = this.validateFieldsPerFieldset();
+
     if (keys(errors).length > 0) {
       this.setState({
         errors,
@@ -645,6 +991,9 @@ class Form extends Component {
                       value={this.state.formData[field]}
                       required={schema.required.indexOf(field) !== -1}
                       onChange={this.onChangeField}
+                      onBlur={this.onBlurField}
+                      onClick={this.onClickInput}
+                      dateOnly={schema.properties[field].widget === 'date'}
                       key={field}
                       error={this.state.errors[field]}
                     />
@@ -671,6 +1020,8 @@ class Form extends Component {
                   tabular: true,
                   className: 'formtabs',
                 }}
+                onTabChange={this.onTabChange}
+                activeIndex={this.state.activeIndex}
                 panes={map(schema.fieldsets, item => ({
                   menuItem: item.title,
                   render: () => [
@@ -688,6 +1039,9 @@ class Form extends Component {
                         value={this.state.formData[field]}
                         required={schema.required.indexOf(field) !== -1}
                         onChange={this.onChangeField}
+                        onBlur={this.onBlurField}
+                        onClick={this.onClickInput}
+                        dateOnly={schema.properties[field].widget === 'date'}
                         key={field}
                         error={this.state.errors[field]}
                       />
@@ -731,6 +1085,9 @@ class Form extends Component {
                     value={this.state.formData[field]}
                     required={schema.required.indexOf(field) !== -1}
                     onChange={this.onChangeField}
+                    onBlur={this.onBlurField}
+                    onClick={this.onClickInput}
+                    dateOnly={schema.properties[field].widget === 'date'}
                     key={field}
                     error={this.state.errors[field]}
                   />