+++ src/components/manage/Widgets/EmailWidget.jsx
@@ -1,6 +1,6 @@
 /**
- * PasswordWidget component.
- * @module components/manage/Widgets/PassswordWidget
+ * EmailWidget component.
+ * @module components/manage/Widgets/EmailWidget
  */
 
 import React from 'react';
@@ -8,16 +8,15 @@ import PropTypes from 'prop-types';
 import { Form, Grid, Input, Label } from 'semantic-ui-react';
 import { map } from 'lodash';
 
-/**
- * PasswordWidget component class.
- * @function PasswordWidget
- * @returns {string} Markup of the component.
+/** EmailWidget function component
+ * @function EmailWidget
+ * @returns {string} Markup of the component
  */
-const PasswordWidget = ({
+const EmailWidget = ({
   id,
   title,
-  required,
   description,
+  required,
   error,
   value,
   onChange,
@@ -54,7 +53,7 @@ const PasswordWidget = ({
             <Input
               id={inputId}
               name={id}
-              type="password"
+              type="email"
               value={value || ''}
               onChange={({ target }) =>
                 onChange(id, target.value === '' ? undefined : target.value)
@@ -86,11 +85,11 @@ const PasswordWidget = ({
 };
 
 /**
- * Property types.
+ * Property types
  * @property {Object} propTypes Property types.
  * @static
  */
-PasswordWidget.propTypes = {
+EmailWidget.propTypes = {
   id: PropTypes.string.isRequired,
   title: PropTypes.string.isRequired,
   description: PropTypes.string,
@@ -107,7 +106,7 @@ PasswordWidget.propTypes = {
  * @property {Object} defaultProps Default properties.
  * @static
  */
-PasswordWidget.defaultProps = {
+EmailWidget.defaultProps = {
   description: null,
   required: false,
   error: [],
@@ -116,4 +115,4 @@ PasswordWidget.defaultProps = {
   maxLength: null,
 };
 
-export default PasswordWidget;
+export default EmailWidget;