+++ src/config/Widgets.jsx
@@ -15,6 +15,8 @@ import TextWidget from '@plone/volto/components/manage/Widgets/TextWidget';
 import TokenWidget from '@plone/volto/components/manage/Widgets/TokenWidget';
 import WysiwygWidget from '@plone/volto/components/manage/Widgets/WysiwygWidget';
 import ReferenceWidget from '@plone/volto/components/manage/Widgets/ReferenceWidget';
+import UrlWidget from '@plone/volto/components/manage/Widgets/UrlWidget';
+import EmailWidget from '@plone/volto/components/manage/Widgets/EmailWidget';
 
 export const DatetimeWidget = loadable(() =>
   import('@plone/volto/components/manage/Widgets/DatetimeWidget'),
@@ -31,8 +33,11 @@ export const widgetMapping = {
     richtext: WysiwygWidget,
     textarea: TextareaWidget,
     datetime: DatetimeWidget,
+    date: DatetimeWidget,
     password: PasswordWidget,
     file: FileWidget,
+    url: UrlWidget,
+    email: EmailWidget,
   },
   vocabulary: {
     'plone.app.vocabularies.Catalog': ReferenceWidget,
@@ -43,7 +48,10 @@ export const widgetMapping = {
     array: ArrayWidget,
     object: FileWidget,
     datetime: DatetimeWidget,
+    date: DatetimeWidget,
     password: PasswordWidget,
+    url: UrlWidget,
+    email: EmailWidget,
   },
 };